🐛 now works

+++ sa.h
@@ -6,7 +6,7 @@ using namespace std;
 class Sa{
 public:
     
-    Sa(int initialTemperature, double coolingRate);
+    Sa(double initialTemperature, double coolingRate);
     void mainSaFullRandom(Board board, Board auxBoard);
     void mainSaOptimized(Board board, Board auxBoard);
     void drawNeighbor(Board& auxBoard);
